requirements.txt
kafka-python==2.0.2
requests==2.25.1
psycopg2-binary==2.8.6

pytest==6.2.2
pytest-cov==2.11.1
responses==0.12.1
mypy==0.800
